refactor: move script-file into fs directory
