+++ packages/eslint-checker/src/fs/script-file.ts
